shop/requirements.txt
amqp==5.1.1
asgiref==3.7.1
async-timeout==4.0.2
beautifulsoup4==4.12.2
billiard==3.6.4.0
bs4==0.0.1
celery==5.2.7
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
Django==4.2.1
django-admin-interface==0.26.0
django-colorfield==0.8.0
django-jalali==6.0.1
django-jalali-date==1.0.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
jdatetime==4.1.1
kombu==5.2.4
Pillow==9.5.0
prompt-toolkit==3.0.38
psycopg==3.1.9
PyJWT==2.7.0
python-slugify==8.0.1
pytz==2023.3
redis==4.5.5
six==1.16.0
soupsieve==2.4.1
sqlparse==0.4.4
text-unidecode==1.3
typing_extensions==4.6.3
vine==5.0.0
wcwidth==0.2.6
